logger/views.py
from rest_framework.decorators import api_view
from rest_framework.response import Response
from rest_framework import status
from django.contrib.auth.models import User
from django.shortcuts import get_object_or_404
from datetime import datetime
from jwt import ExpiredSignatureError, InvalidTokenError
from django.http import HttpResponse, HttpRequest

