requirements.txt
asgiref==3.8.1
certifi==2024.7.4
channels==4.1.0
charset-normalizer==3.3.2
colorama==0.4.6
cors==1.0.1
Django==5.1.2
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
dnspython==2.7.0
environ==1.0
filelock==3.15.4
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.7
pika==1.3.2
postgres==4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg2-pool==1.2
PyJWT==2.9.0
pymongo==4.10.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
tldextract==5.1.2
typing_extensions==4.12.2
urllib3==2.2.2
zope.event==5.0
zope.interface==7.0.1
